+++ crates/jsonschema/src/compiler.rs
@@ -951,8 +951,11 @@ fn compile_without_cache<'a>(
                 // Check if this keyword is overridden, then check the standard definitions
                 if let Some(factory) = ctx.get_keyword_factory(keyword) {
                     let path = ctx.location().join(keyword);
-                    let validator =
-                        CustomKeyword::new(factory.init(schema, value, path.clone())?, path);
+                    let validator = CustomKeyword::new(
+                        factory.init(schema, value, path.clone(), keyword)?,
+                        path,
+                        keyword.clone(),
+                    );
                     let validator: BoxedValidator = Box::new(validator);
                     validators.push((Keyword::custom(keyword), validator));
                 } else if let Some((keyword, validator)) = keywords::get_for_draft(ctx, keyword)